THOT dimensions: uint32, THOT256 wallet key, THOT65536 + THOT1048576 quantum
All THOT contracts upgraded from uint8 → uint32 dimensions for full expansion:
- THOT.sol, THINK.sol, tNFT.sol, TransmuteAgent.sol, iNFT.sol

Modular _isValidDimension() with 11 standard dimensions:
  THOT8       — root of THOT, the seed dimension
  THOT64      — lightweight vectors
  THOT256     — wallet key dimension (32-byte key = 256 bits)
  THOT512     — standard 8x8x8 3D knowledge clusters
  THOT768     — high-fidelity optimized tensors
  THOT1024    — embedding-native (mxbai-embed-large)
  THOT2048    — cypherpunk2048 high-capacity
  THOT4096    — quantum-aware tensor space
  THOT8192    — quantum-aware high-dimensional
  THOT65536   — theoretical quantum-resistant (2^16)
  THOT1048576 — theoretical post-quantum (2^20)

uint32 supports up to 4,294,967,295 — future dimensions added by extending
_isValidDimension() only. No other code changes needed.

iNFT UI updated with all 11 dimension options.

+++ daio/contracts/THOT/agents/TransmuteAgent.sol
@@ -64,7 +64,7 @@ contract TransmuteAgent is Ownable {
     function transmuteDataFull(
         address recipient,
         string memory inputData,
-        uint8 dimensions,
+        uint32 dimensions,
         uint8 parallelUnits,
         string memory metadataURI
     ) external onlyOwner returns (uint256) {
